+++ exports/internal.go
@@ -7,23 +7,19 @@ SPDX-License-Identifier: Apache-2.0
 package exports
 
 import (
-	"context"
 	"net/http"
 
 	"github.com/go-chi/chi/v5"
-	"github.com/redhatinsights/export-service-go/errors"
-)
 
-// geturlParams fetches the urlParams from the context.
-func geturlParams(ctx context.Context) *urlParams {
-	return ctx.Value(urlParamsKey).(*urlParams)
-}
+	"github.com/redhatinsights/export-service-go/errors"
+	"github.com/redhatinsights/export-service-go/middleware"
+)
 
 // InternalRouter is a router for all of the internal routes which require exportuuid,
 // application name, and resourceuuid.
 func InternalRouter(r chi.Router) {
 	r.Route("/{exportUUID}/{application}/{resourceUUID}", func(sub chi.Router) {
-		sub.Use(URLParams)
+		sub.Use(middleware.URLParamsCtx)
 		sub.Post("/upload", PostUpload)
 		sub.Post("/error", PostError)
 	})
@@ -32,7 +28,7 @@ func InternalRouter(r chi.Router) {
 // PostError receives a POST request from the export source which contains the
 // errors associated with creating the export for the given resource.
 func PostError(w http.ResponseWriter, r *http.Request) {
-	params := geturlParams(r.Context())
+	params := middleware.GetURLParams(r.Context())
 	if params == nil {
 		errors.InternalServerError(w, "unable to parse url params")
 		return
@@ -43,7 +39,7 @@ func PostError(w http.ResponseWriter, r *http.Request) {
 // PostUpload receives a POST request from the export source containing
 // the exported data. This data is uploaded to S3.
 func PostUpload(w http.ResponseWriter, r *http.Request) {
-	params := geturlParams(r.Context())
+	params := middleware.GetURLParams(r.Context())
 	if params == nil {
 		errors.InternalServerError(w, "unable to parse url params")
 		return
